requirements.txt
Django==1.8.5
django-allauth==0.23.0
djangorestframework==3.2.4
facepy==1.0.7
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
MySQL-python==1.2.5
oauth2client==1.5.1
oauthlib==1.0.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
rsa==3.2
simplejson==3.8.0
six==1.10.0
uritemplate==0.6
wheel==0.24.0
dj-database-url==0.3.0
dj-static==0.0.6
